implemented linting, changed file locations, commenting
